test: split durable inventory fallback coverage
